+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\ClassController;
 use App\Http\Controllers\ExtracurricularController;
 use App\Http\Controllers\StudentController;
@@ -18,30 +19,29 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::get('/', function () {
-    return view('home', [
-        'name' => 'Nasruddin',
-        'role' => 'Admin',
-        'buah' => ['Pisang', 'Apel', 'Jeruk', 'Semangka', 'Nanas']
-    ]);
-});
+    return view('home');
+})->middleware('auth');
 
+Route::get('/login', [AuthController::class, 'login'])->name('login')->middleware('guest');
+Route::post('/login', [AuthController::class, 'authenticate'])->middleware('guest');
+Route::get('/logout', [AuthController::class, 'logout'])->middleware('auth');
 
-Route::get('/students', [StudentController::class, 'index']);
-Route::get('/student-detail/{id}', [StudentController::class, 'show']);
-Route::get('/student-add', [StudentController::class, 'create']);
-Route::post('/student', [StudentController::class, 'store']);
-Route::get('/student-edit/{id}', [StudentController::class, 'edit']);
-Route::put('/student/{id}', [StudentController::class, 'update']);
-Route::get('/student-delete/{id}', [StudentController::class, 'delete']);
-Route::delete('/student-destroy/{id}', [StudentController::class, 'destroy']);
-Route::get('/students-deleted', [StudentController::class, 'deletedStudent']);
-Route::get('/student/{id}/restore', [StudentController::class, 'restore']);
+Route::get('/students', [StudentController::class, 'index'])->middleware('auth');
+Route::get('/student-detail/{id}', [StudentController::class, 'show'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::get('/student-add', [StudentController::class, 'create'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::post('/student', [StudentController::class, 'store'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::get('/student-edit/{id}', [StudentController::class, 'edit'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::put('/student/{id}', [StudentController::class, 'update'])->middleware(['auth', 'must-admin-or-teacher']);
+Route::get('/student-delete/{id}', [StudentController::class, 'delete'])->middleware(['auth', 'must-admin']);
+Route::delete('/student-destroy/{id}', [StudentController::class, 'destroy'])->middleware(['auth', 'must-admin']);
+Route::get('/students-deleted', [StudentController::class, 'deletedStudent'])->middleware(['auth', 'must-admin']);
+Route::get('/student/{id}/restore', [StudentController::class, 'restore'])->middleware(['auth', 'must-admin-or-teacher']);
 
-Route::get('/class', [ClassController::class, 'index']);
-Route::get('/class-detail/{id}', [ClassController::class, 'show']);
+Route::get('/class', [ClassController::class, 'index'])->middleware('auth');;
+Route::get('/class-detail/{id}', [ClassController::class, 'show'])->middleware('auth');
 
-Route::get('/ektracurricular', [ExtracurricularController::class, 'index']);
-Route::get('/ektracurricular-detail/{id}', [ExtracurricularController::class, 'show']);
+Route::get('/ektracurricular', [ExtracurricularController::class, 'index'])->middleware('auth');;
+Route::get('/ektracurricular-detail/{id}', [ExtracurricularController::class, 'show'])->middleware('auth');
 
-Route::get('/teacher', [TeacherController::class, 'index']);
-Route::get('/teacher-detail/{id}', [TeacherController::class, 'show']);
+Route::get('/teacher', [TeacherController::class, 'index'])->middleware('auth');;
+Route::get('/teacher-detail/{id}', [TeacherController::class, 'show'])->middleware('auth');
